Forgotten password function complete.
The forgotten password page now prompts the user for an email address and displays a captcha.  Once past the captcha the database is checked to see if the email address is valid.  If so the account status is set to disabled, a new random password is generated, and an email is sent to the user.  This email provides a link to the reset_account page which is based on the original activate_account page, but only enables the account.  Closes #48.  Closes #42.